fix for incorrect escaping of the redirect header's parameters.
Commit 08a27dcfb5 introduced use of the PHP "htmlentities" function
to rectify some XSS vulnerabilities. In doing so, the redirect sent
back to the client was mis-escaped so that the '&' parameter separator
was represented as '&amp;'. This resulted in the second and following
parameter's names being prefixed with 'amp;' and that stopped the
login process from working.

This commit removes the call to 'htmlentities' on the requested URL
and replaces it with a call where the request is applied to a <link>
tag. This way, it's escaped in the HTML but not in the redirect header.

Works with PHP version 5.6.8.